[SPARK-11646] WholeTextFileRDD should return Text rather than String
